// src/types/api.ts
export interface PaginatedResult<T> {
  data: T[];
  meta: {
    total: number;
    lastPage: number;
    currentPage: number;
    perPage: number;
    prev: number | null;
    next: number | null;
  };
}

export interface BaseServerException {
  message: string;
  code: number;
  status: number;
}